pyproject.toml
[project]
name = "gemma-env"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.1",
    "requests>=2.32.5",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "transformers>=5.7.0",
]
